pipeline/watch_events/pyproject.toml
[project]
name = "shift-backend"
version = "0.1.0"
description = "SHIFT Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "httpx==0.25.2",
    "python-jose[cryptography]==3.3.0",
    "pydantic==2.5.0",
    "cachetools==5.3.2",
    "google-cloud-firestore",
    "google-cloud-bigquery",
    "google-cloud-pubsub",
]

